+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             android:name=".ui.settings.activity.SettingsActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.player.AudioPlayerActivity"
+            android:name=".ui.player.activity.AudioPlayerActivity"
             android:exported="false" />
         <activity
             android:name=".ui.main.MainActivity"
